project/requirements.txt
Flask==2.0.2
Werkzeug==2.2.2
Flask-PyMongo==2.3.0

